sfED

TODO: general information

Index

sfED.Eigenspace

sfED.Fockstates

Base.:==

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

sfED.getCdagmatrix

sfED.getCmatrix

sfED.getCsign

sfED.getEps

sfED.getEvalEvecs

sfED.getEvalveclist

sfED.getFuckingLarge2partTerm

sfED.getG0

sfED.getGF

sfED.getGF2part

sfED.getHamiltonian

sfED.getHopping

sfED.getMatrixElem1Particle

sfED.getMatrixElem2Particle

sfED.getNmaxFromAllstates

sfED.getPossibleTransitions

sfED.getSigma

sfED.getSpin

sfED.getTmatrix

sfED.getUJmatrix

sfED.getUdensity

sfED.getUnondensity

sfED.getZ

sfED.indexSpinConfig

sfED.noSpinConfig

sfED.nstatesNS

sfED.nstatesNS

sfED.spinConfig

sfED.writeEvalContributions

sfED.writeEvalContributionsSectors

sfED.writeEvalInfo

sfED.writeGF

sfED.writeGF2part

sfED.writeGFnorm

sfED.writeMatrixGnuplot

sfED.writeStateInfo

sfED.writeTransitionsOverlaps

sfED.Eigenspace — Method

Eigenspace(eps,tmatrix,Umatrix,Jmatrix,pSimulation.mu,fockstates,pNumerics) Constructor for the Eigenspace struct

sfED.Fockstates — Method

Fockstates(norb) Constructor for the Fockstates struct

Base.:== — Method

==(Transition,Transition)

Base.getindex — Method

getindex(es,n,is,:) Return the Eigenvalues and -states in the N=n,S=S(is) subspace

Base.getindex — Method

getindex(es,n,is,i) Return the i-th Eigenvalue and Eigenvector in the N=n,S=S(is) subspace

Base.getindex — Method

getindex(es,i) Just return the i-th eigenvalue and state

Base.getindex — Method

getindex(fs,n,is,:) Return the Fock states in the N=n,S=S(is) subspace

Base.getindex — Method

getindex(fs,n,is,i) Return the i-th Fock state in the N=n,S=S(is) subspace

Base.getindex — Method

getindex(fs,i) Just return the i-th Fock state in fs

sfED.getCdagmatrix — Method

getCdagmatrix(crea,subspace1,subspace2)

Generate the matrix for the creation operator acting on the subspace with N-1,S-1 quantum numbers we act subspace1 * cdag * subspace2 subspace 2 has all basis states with N-1,S-1 because we have previously annihilated one up electron subspace 1 has all basis states with N,S now just determine the transformation when acting the creation operator crea on it crea is the orbital/spin index

sfED.getCmatrix — Method

getCmatrix(anni::Int64,subspace1, subspace2)

Generate the matrix for the annihilation operator acting on the subspace with N,S quantum number we act subspace1 * c * subspace2 subspace 2 has all basis states with N,S subspace 1 has all basis states with N-1,S-dS now just determine the transformation when acting the annihilation operator anni on it anni is the orbital/spin index

sfED.getCsign — Method

getCsign(i,state)

return the sign when creating/annihilating an electron at position i in state.

sfED.getEps — Method

getEps(pNumerics,pModel)

Return the local orbital levels, define them by hand here. Values are perturbed by smallValue TODO: this should be overloaded to accept text file or command line inputs.

sfED.getEvalEvecs — Method

getEvalEvecs(Hamiltonian)

Diagonalize a given Hamiltonian. Eigenvalues will be cast to real, since the Hamiltonian is hermitian.

sfED.getEvalveclist — Method

getEvalveclist(eps,tmatrix,Umatrix,Jmatrix,allstates)

Create the N,S submatrices of the Hamiltonian, solve it and return the Eigenvalues and Vectors in a List

sfED.getFuckingLarge2partTerm — Method

getFuckingLarge2partTerm(w1,w2,w3,Em,En,Eo,Ep,beta)

Evaluate the Lehmann term for the 2part Green's function

sfED.getG0 — Method

getG0(eps, tmatrix, w)

Construct the noninteracting Green's function from onsite energy eps, hoppings between sites i and j, tmatrix and Matsubara grid w.

sfED.getGF — Method

getGF(evallist, eveclist, allstates, pModel, pSimulation, pFreq, pNumerics)

Evaluation of the Lehmann representation for the interacting finite-temperature Green's function We sum over all Eigenstates evallist with electron number N and all other eigenstates n2 with electron number N-1 Since we have spin degeneracy, we only calculate the up GF, so we annihilate one up spin, so we also only sum over all S-1 states for given S(n1)

sfED.getGF2part — Method

getGF2part(evallist, eveclist, allstates, pModel, pSimulation, pFreq, pNumerics)

Evaluation of the Lehmann representation for the interacting finite-temperature 2-particle Green's function We sum over all Eigenstates evallist with electron number N and all other eigenstates connected by c and c^dagger We choose the definition G^(2)up,dn = <T c^dagup(t1) cup(t2) c^dagdn(t3) c_dn(0) >

sfED.getHamiltonian — Method

getHamiltonian(eps,tmatrix,Umatrix,Jmatrix,states)

Constructs hamiltonian from onsite energies eps, hoppings tmatrix, coulomb interactions Umatrix and Jmatrix over precomputed states.

sfED.getHopping — Method

getHopping(istate, jstate, tmatrix)

return the hopping contribution for states <i| |j>

sfED.getMatrixElem1Particle — Method

getMatrixElem1Particle(c1, a1, istate, jstate)

Calculate the matrix element for a one particle operator given the creation and annihilation operator with indices c1 and a1 repsectively.

sfED.getMatrixElem2Particle — Method

getMatrixElem2Particle(c1,c2,a3,a4,istate,jstate)

Calculate the matrix element for a two particle operator given the creation and annihilation operators with indices c1, c2 and a3,a4.

sfED.getNmaxFromAllstates — Method

getNmaxFromAllstates(allstates)

return the maximum number of electrons possible determined from the allstates array

sfED.getPossibleTransitions — Method

getPossibleTransitions(eigenspace,fockstates,flavors,pNumerics,expCutoff)

Calculate the overlap elements between all Eigenstates acting on c/c^dagger specified by flavors. The function returns the overlap array and possibleTransitions for cdagger,c for the flavors specified in the flavors array (1=orb1,up, 2=orb1,dn, 3=orb2,up, ....)

sfED.getSigma — Method

getSigma(G0, G)

Calculate the Selfenergy from G0 and G

sfED.getSpin — Method

getSpin(states)

total spin S of state.

sfED.getTmatrix — Method

getTmatrix(pModel)

Return the hoppingmatrix, defined by hand here TODO: this should be overloaded to accept text file or command line inputs.

sfED.getUJmatrix — Method

getUJmatrix(pModel)

Return the Coulomb interaction matrices, defined by hand here TODO: this should be overloaded to accept text file or command line inputs.

sfED.getUdensity — Method

getUdensity(state, Umatrx, Jmatrix)

return the density density part of the Coulomb interaction contribution

sfED.getUnondensity — Method

getUnondensity(istate, jstate, Jmatrix)

return the spin-flip and pair-hopping part of the interaction contribution

sfED.getZ — Method

getZ(evallist)

Calculate partition function from the eigenvalues in evallist.

sfED.indexSpinConfig — Method

indexSpinConfig(S,n,Nmax)

Index i corresponding to a spin config S for given n and nMax

Example

julia> indexSpinConfig(spinConfig(2,2,4),2,4)
2

sfED.noSpinConfig — Method

noSpinConfig(n, Nmax)

Number of configurations with different total spin S possible for given n and Nmax.

Examples

julia> sfED.noSpinConfig(2,4)
3

sfED.nstatesNS — Method

nstatesNS(es,N,S) Return number of states for given N and S

sfED.nstatesNS — Method

nstatesNS(fs,N,S) Return number of states for given N and S

sfED.spinConfig — Method

spinConfig(i,n,Nmax)

i-th possible total spin value S for given n, nMax. i starts at 1

Examples

julia> sfED.spinConfig(1,2,4)
-2
julia> sfED.spinConfig(2,2,4)
0
julia> sfED.spinConfig(3,2,4)
2

sfED.writeEvalContributions — Method

writeEvalContributions(file, evalContributions)

write the weight contribtions of each Eigenstate sorted by Eigenenergy

sfED.writeEvalContributionsSectors — Method

writeEvalContributionsSectors(filename, evalContributions)

write the weight contribtions of each Eigenstate sorted by N,S quantum numbers

sfED.writeEvalInfo — Method

writeEvalInfo(evallist, eveclist, allstates)

Write the Eigenvalues and some info

sfED.writeGF — Method

writeGF(filename, G, w)

Writes a Green's function G(norb,norb,nw) in human readable (fixed column width format) form into a file. mgrid is the associated Matsuabra grid (as indices or evaluated).

sfED.writeGF2part — Method

writeGF2part(filename, G, w)

Writes a Green's function G(nw,nw,nw) in human readable (fixed column width format) form into a file. mgrid is the associated Matsuabra grid (as indices or evaluated).

sfED.writeGFnorm — Method

writeGFnorm(gfdiagnorm)

Write the calculated normalization of the diagonal Green's function elements

sfED.writeMatrixGnuplot — Method

writeMatrixGnuplot(filename, matrix)

Writes a Matrix matrix(norb1,norb2) in human readable (fixed column width format) form into a file. Only writes the real part and writes a linebreak after each row so that Gnuplot sp w pm3d works .

sfED.writeStateInfo — Method

writeStateInfo(allstates)

Write all states and N S quantum numbers

sfED.writeTransitionsOverlaps — Method

writeTransitionsOverlaps(filename,overlaps)

Write out the overlap elements between all Eigenstates acting on c/c^dagger